==> CITE-seq/08-sub_clustering/GCBC/01-GCBC_B_cells_subset_by_cluster.Rmd <==
---
title: "GCBC B cell clusters"
date: "`r format(Sys.time(), '%B %Y')`"
output:
  html_document:
    toc: yes
    df_print: paged
  BiocStyle::html_document:
    toc: yes
    toc_float: yes
    number_sections: yes
---


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(knitr)
setwd("/home/srashmi/Documents/tonsil_atlas_citeseq_vdj_20210505/")
```

# Define Parameters

## Load packages

```{r } 
library(Seurat)
library(flexclust)
library(tidyverse)
library(dplyr)
library(ggplot2)
library(patchwork)
library(scales)
library(harmony)
```


## Define variables and functions

```{r}
path_to_save_obj <- "/home/srashmi/Documents/tonsil_atlas_citeseq_vdj_20210505/results"
path_to_save_citeseq_seurat_obj <- str_c(
  path_to_save_obj,
  "tonsil_cite_seq_annotated_filtered.rds",
  sep = "/"
)
path_to_save_seurat_obj <- str_c(
  "objects/seurat_object_cite_seq_gcbc_positive_object.rds",
  sep = "/"
)
```

## Load data

```{r data}
seurat_obj <- readRDS(path_to_save_citeseq_seurat_obj)
```

## get metadata

```{r get metadata}
metadata <- seurat_obj@meta.data 
```

# Extract GCBC cells

```{r}
GCBC_pos_cells <- rownames(subset(metadata, metadata$annotation == "GCBC"))
GCBC_pos_clusters_obj = subset(seurat_obj,cells = GCBC_pos_cells)
```


## QC of GCBC enriched clusters

```{r}
DimPlot(GCBC_pos_clusters_obj)
```

```{r}
FeaturePlot(GCBC_pos_clusters_obj, features = c("CD19.1","CD3","CD5"), ncol = 2)
```


## Save the seurat object 

```{r}
saveRDS(GCBC_pos_clusters_obj, file = path_to_save_seurat_obj)
```

# Session Information

```{r}
sessionInfo()
```
